Ajout détection trajets Tan
Premiers tests Google maps WayPoints

+++ projetweb/src/main/java/com/projetweb/bean/Coordonnee.java
@@ -20,46 +20,46 @@ public class Coordonnee implements java.io.Serializable{
 	 * Longitude :X
 	 */
 	@SerializedName("lng")
-	private float longitude;
+	private double longitude;
 	
 	/**
 	 * Latitude : Y
 	 */
 	@SerializedName("lat")
-	private float latitude;
+	private double latitude;
 
 	public Coordonnee(){
 	}
 	
-	public Coordonnee(float latitude,float longitude){
+	public Coordonnee(double latitude,double longitude){
 		this.latitude = latitude;
 		this.longitude = longitude;
 	}
 	/**
 	 * @return the longitude
 	 */
-	public float getLongitude() {
+	public double getLongitude() {
 		return longitude;
 	}
 
 	/**
 	 * @param longitude the longitude to set
 	 */
-	public void setLongitude(float longitude) {
+	public void setLongitude(double longitude) {
 		this.longitude = longitude;
 	}
 
 	/**
 	 * @return the latitude
 	 */
-	public float getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
 
 	/**
 	 * @param latitude the latitude to set
 	 */
-	public void setLatitude(float latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
